+++ SDK_dotnet/Error/IError.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Gluwa.Error
+{
+    public interface IError
+    {
+    }
+}
