+++ SDK_dotnet/Error/PublicError.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Net;
+using System.Text;
+
+namespace Gluwa.Error
+{
+    public sealed class PublicError : Error<HttpStatusCode>
+    {
+    }
+}
