+++ SDK_dotnet/Error/Result.cs
@@ -1,6 +1,8 @@
-﻿namespace SDK_dotnet.Utils
+﻿using Gluwa.Error;
+
+namespace Gluwa.Error
 {
-    public sealed class Result<T>
+    public sealed class Result<T, E> where E : IError
     {
         public System.Guid ID { get; } = System.Guid.NewGuid();
         public bool IsSuccess { get; set; }
@@ -9,5 +11,6 @@
             get { return !IsSuccess; }
         }
         public T Data { get; set; }
+        public E Error { get; set; }
     }
 }
