+++ SDK_dotnet/Models/BalanceResponse.cs
@@ -1,13 +1,22 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace SDK_dotnet.Models
+namespace Gluwa.Models
 {
     public sealed class BalanceResponse
     {
+        /// <summary>
+        /// Address balance.
+        /// </summary>
         public string Balance { get; set; }
 
+        /// <summary>
+        /// Currency for the address balance.
+        /// </summary>
         public ECurrency? Currency { get; set; }
 
+        /// <summary>
+        /// Current nonce for the address. For Gluwacoin currencies only.
+        /// </summary>
         [Range(0, ulong.MaxValue)]
         public ulong? Nonce { get; set; }
     }
