+++ SDK_dotnet/Models/FeeResponse.cs
@@ -1,12 +1,18 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace SDK_dotnet.Models
+namespace Gluwa.Models
 {
     public sealed class FeeResponse
     {
+        /// <summary>
+        /// Currency that the fee is for.
+        /// </summary>
         [Required]
         public ECurrency? Currency { get; set; }
-    
+        
+        /// <summary>
+        /// Current minimum amount of the fee that Gluwa will accept.
+        /// </summary>
         [Required]
         public string MinimumFee { get; set; }
     }
