+++ SDK_dotnet/Models/PayLoad.cs
@@ -1,4 +1,4 @@
-﻿namespace SDK_dotnet.Models
+﻿namespace Gluwa.Models
 {
     public sealed class PayLoad
     {
