+++ SDK_dotnet/Models/TransactionRequest.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.ComponentModel.DataAnnotations;
 
-namespace SDK_dotnet.Models
+namespace Gluwa.Models
 {
     internal sealed class TransactionRequest
     {
